requirements.txt
numpy
pandas
opencv-python
scikit-learn
joblib
pyyaml
tqdm
ultralytics
